requirements.txt
attrs==23.1.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.3
GeoAlchemy2==0.12.0
geopandas==0.12.2
greenlet==2.0.2
importlib-metadata==6.6.0
munch==2.5.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
psycopg2-binary==2.9.6
pyproj==3.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
shapely==2.0.1
six==1.16.0
SQLAlchemy==1.4.48
typing-extensions==4.5.0
tzdata==2023.3
zipp==3.15.0
